refactor: ghost
