+++ backend/component/ghost/config.go
@@ -1,0 +1,175 @@
+package ghost
+
+import (
+	"fmt"
+	"strconv"
+	"strings"
+
+	"github.com/github/gh-ost/go/base"
+	ghostsql "github.com/github/gh-ost/go/sql"
+	"github.com/pkg/errors"
+
+	"github.com/bytebase/bytebase/backend/common"
+	"github.com/bytebase/bytebase/backend/store"
+)
+
+// Config is the configuration for gh-ost migration.
+type Config struct {
+	// serverID should be unique
+	ServerID             uint
+	Host                 string
+	Port                 string
+	User                 string
+	Password             string
+	Database             string
+	Table                string
+	AlterStatement       string
+	SocketFilename       string
+	PostponeFlagFilename string
+	Noop                 bool
+
+	// vendor related
+	IsAWS bool
+}
+
+// GetGhostConfig returns a gh-ost configuration for migration.
+func GetGhostConfig(taskID int, database *store.DatabaseMessage, dataSource *store.DataSourceMessage, secret string, instanceUsers []*store.InstanceUserMessage, tableName string, statement string, noop bool, serverIDOffset uint) (Config, error) {
+	var isAWS bool
+	for _, user := range instanceUsers {
+		if user.Name == "'rdsadmin'@'localhost'" && strings.Contains(user.Grant, "SUPER") {
+			isAWS = true
+			break
+		}
+	}
+	password, err := common.Unobfuscate(dataSource.ObfuscatedPassword, secret)
+	if err != nil {
+		return Config{}, err
+	}
+	return Config{
+		Host:                 dataSource.Host,
+		Port:                 dataSource.Port,
+		User:                 dataSource.Username,
+		Password:             password,
+		Database:             database.DatabaseName,
+		Table:                tableName,
+		AlterStatement:       statement,
+		SocketFilename:       getSocketFilename(taskID, database.UID, database.DatabaseName, tableName),
+		PostponeFlagFilename: GetPostponeFlagFilename(taskID, database.UID, database.DatabaseName, tableName),
+		Noop:                 noop,
+		// On the source and each replica, you must set the server_id system variable to establish a unique replication ID. For each server, you should pick a unique positive integer in the range from 1 to 2^32 − 1, and each ID must be different from every other ID in use by any other source or replica in the replication topology. Example: server-id=3.
+		// https://dev.mysql.com/doc/refman/5.7/en/replication-options-source.html
+		// Here we use serverID = offset + task.ID to avoid potential conflicts.
+		ServerID: serverIDOffset + uint(taskID),
+		// https://github.com/github/gh-ost/blob/master/doc/rds.md
+		IsAWS: isAWS,
+	}, nil
+}
+
+func getSocketFilename(taskID int, databaseID int, databaseName string, tableName string) string {
+	return fmt.Sprintf("/tmp/gh-ost.%v.%v.%v.%v.sock", taskID, databaseID, databaseName, tableName)
+}
+
+// GetPostponeFlagFilename gets the postpone flag filename for gh-ost.
+func GetPostponeFlagFilename(taskID int, databaseID int, databaseName string, tableName string) string {
+	return fmt.Sprintf("/tmp/gh-ost.%v.%v.%v.%v.postponeFlag", taskID, databaseID, databaseName, tableName)
+}
+
+// NewMigrationContext is the context for gh-ost migration.
+func NewMigrationContext(config Config) (*base.MigrationContext, error) {
+	const (
+		allowedRunningOnMaster              = true
+		concurrentCountTableRows            = true
+		timestampAllTable                   = true
+		hooksStatusIntervalSec              = 60
+		heartbeatIntervalMilliseconds       = 100
+		niceRatio                           = 0
+		chunkSize                           = 1000
+		dmlBatchSize                        = 10
+		maxLagMillisecondsThrottleThreshold = 1500
+		defaultNumRetries                   = 60
+		cutoverLockTimeoutSeconds           = 60
+		exponentialBackoffMaxInterval       = 64
+		throttleHTTPIntervalMillis          = 100
+		throttleHTTPTimeoutMillis           = 1000
+	)
+	statement := strings.Join(strings.Fields(config.AlterStatement), " ")
+	migrationContext := base.NewMigrationContext()
+	migrationContext.InspectorConnectionConfig.Key.Hostname = config.Host
+	port := 3306
+	if config.Port != "" {
+		configPort, err := strconv.Atoi(config.Port)
+		if err != nil {
+			return nil, errors.Wrap(err, "failed to convert port from string to int")
+		}
+		port = configPort
+	}
+	migrationContext.InspectorConnectionConfig.Key.Port = port
+	migrationContext.CliUser = config.User
+	migrationContext.CliPassword = config.Password
+	migrationContext.DatabaseName = config.Database
+	migrationContext.OriginalTableName = config.Table
+	migrationContext.AlterStatement = statement
+	migrationContext.Noop = config.Noop
+	migrationContext.ReplicaServerId = config.ServerID
+	if config.IsAWS {
+		migrationContext.AssumeRBR = true
+	}
+	// set defaults
+	if err := migrationContext.SetConnectionConfig(""); err != nil {
+		return nil, err
+	}
+	migrationContext.AllowedRunningOnMaster = allowedRunningOnMaster
+	migrationContext.ConcurrentCountTableRows = concurrentCountTableRows
+	migrationContext.HooksStatusIntervalSec = hooksStatusIntervalSec
+	migrationContext.CutOverType = base.CutOverAtomic
+	migrationContext.ThrottleHTTPIntervalMillis = throttleHTTPIntervalMillis
+	migrationContext.ThrottleHTTPTimeoutMillis = throttleHTTPTimeoutMillis
+
+	if migrationContext.AlterStatement == "" {
+		return nil, errors.Errorf("alterStatement must be provided and must not be empty")
+	}
+	parser := ghostsql.NewParserFromAlterStatement(migrationContext.AlterStatement)
+	migrationContext.AlterStatementOptions = parser.GetAlterStatementOptions()
+
+	if migrationContext.DatabaseName == "" {
+		if !parser.HasExplicitSchema() {
+			return nil, errors.Errorf("database must be provided and database name must not be empty, or alterStatement must specify database name")
+		}
+		migrationContext.DatabaseName = parser.GetExplicitSchema()
+	}
+	if migrationContext.OriginalTableName == "" {
+		if !parser.HasExplicitTable() {
+			return nil, errors.Errorf("table must be provided and table name must not be empty, or alterStatement must specify table name")
+		}
+		migrationContext.OriginalTableName = parser.GetExplicitTable()
+	}
+	migrationContext.ServeSocketFile = config.SocketFilename
+	migrationContext.PostponeCutOverFlagFile = config.PostponeFlagFilename
+	migrationContext.TimestampAllTable = timestampAllTable
+	migrationContext.SetHeartbeatIntervalMilliseconds(heartbeatIntervalMilliseconds)
+	migrationContext.SetNiceRatio(niceRatio)
+	migrationContext.SetChunkSize(chunkSize)
+	migrationContext.SetDMLBatchSize(dmlBatchSize)
+	migrationContext.SetMaxLagMillisecondsThrottleThreshold(maxLagMillisecondsThrottleThreshold)
+	migrationContext.SetDefaultNumRetries(defaultNumRetries)
+	migrationContext.ApplyCredentials()
+	if err := migrationContext.SetCutOverLockTimeoutSeconds(cutoverLockTimeoutSeconds); err != nil {
+		return nil, err
+	}
+	if err := migrationContext.SetExponentialBackoffMaxInterval(exponentialBackoffMaxInterval); err != nil {
+		return nil, err
+	}
+	return migrationContext, nil
+}
+
+// GetTableNameFromStatement gets the table name from statement for gh-ost.
+func GetTableNameFromStatement(statement string) (string, error) {
+	// Trim the statement for the parser.
+	// This in effect removes all leading and trailing spaces, substitute multiple spaces with one.
+	statement = strings.Join(strings.Fields(statement), " ")
+	parser := ghostsql.NewParserFromAlterStatement(statement)
+	if !parser.HasExplicitTable() {
+		return "", errors.Errorf("failed to parse table name from statement, statement: %v", statement)
+	}
+	return parser.GetExplicitTable(), nil
+}
